move sleep state
